style: Renaming the authentication exceptions module
